make Swup export utils and helpers from the main
