+++ src/modules/enterPage.ts
@@ -1,5 +1,5 @@
 import { nextTick } from '../utils.js';
-import Swup from '../index.js';
+import Swup from '../Swup.js';
 
 const enterPage = function(
 	this: Swup,
