+++ src/modules/on.ts
@@ -1,4 +1,4 @@
-import Swup from '../index.js';
+import Swup from '../Swup.js';
 
 export type EventType =
 	| 'animationInDone'
