+++ src/modules/renderPage.ts
@@ -1,5 +1,5 @@
 import { Location, updateHistoryRecord, getCurrentUrl } from '../helpers.js';
-import Swup from '../index.js';
+import Swup from '../Swup.js';
 import { PageRecord } from './Cache.js';
 
 const renderPage = function(
